+++ docs/Changelog.md
@@ -32,6 +32,7 @@ Nussknacker versions
   * akka-http 10.1 -> 10.2
   * akka-http-circe 1.28 -> 1.38
 * [#2684](https://github.com/TouK/nussknacker/pull/2684) Handled 'Restarting' state in Embedded DeploymentManager when the embedded scenario is failing
+* [#2686](https://github.com/TouK/nussknacker/pull/2686) Rename `ServiceWithStaticParameters` to `EagerServiceWithStaticParameters` to avoid confusion about lazy and eager parameters used by default
 
 1.1.1 (Not released yet)
 --------------------